Version 2.99995b”

+++ src/it/unitn/ing/rista/diffr/SizeStrainSymModel.java
@@ -82,7 +82,7 @@ public class SizeStrainSymModel extends XRDcat {
 
   double cryststrain[] = new double[2];
 
-  public double[] getCrystalliteMicrostrain(double d_space, int h, int k, int l, double[] texture_angles) {
+  public double[] getCrystalliteMicrostrain(double d_space, double h, double k, double l, double[] texture_angles) {
 
 //		cryststrain[0] = 0.0;
 //		cryststrain[1] = 0.0;
